Add files via upload
This repositary contains codes for time series analysis and models. Starting with exploratory time series data analysis, moving on to simple models such as AR and MA models and finally ending with estimation of complex models such as ARIMAX, SARIMAX, ARCH and GARCH.  A simple time series forecasting business case study is also included in the repositary along with input data used for this analysis.